user_id==token

+++ back-end/app/controllers/article.controller.js
@@ -1,35 +1,30 @@
 
 const express = require("express");
 const app = express();
-const config = require('../config/auth.config');
+// const config = require('../config/auth.config');
 const db = require("../models");
 const Op = db.Sequelize.Op;
 const Article = db.article;
-const Comment = db.comment;
 const User = db.user;
 const jwt = require("jsonwebtoken");
 const fileUpload = require('express-fileupload');
 const fs = require("fs")
-const path = require('path');
 app.use(fileUpload());
 
 // Create and Save a new Article  ::::::::::::::::::::::::::::
-exports.create = (req, res) => {
+exports.create = async (req, res) => {
 
-  if (!req.files || Object.keys(req.files).length === 0) {
-    return res.status(401).send('No files were uploaded.');
-  }
-
-  const imageFile = `${req.protocol}://${req.get('host')}/ressources/${req.file.filename}`;
+  const imageFile = `${req.protocol}://${req.get('host')}/static/assets/uploads/${req.file.filename}`;
+  // const imageFile = req.file.filename;
   const token = req.headers['x-access-token'];
-  const decoded = jwt.decode(token);
-  console.log(imageFile)
+  const decoded = await jwt.decode(token);
   const article = {
     title: req.body.title,
     description: req.body.description,
     username: req.body.username,
     user_id: decoded.id,
     image: imageFile
+
   };
   // Save Article in the database
   Article.create(article)
@@ -39,10 +34,11 @@ exports.create = (req, res) => {
     .catch(err => {
       res.status(500).send({
         message:
-          err.message || "Some error occurred while creating the Article."
+          err.message || "Une erreur s'est produite lors de la creation de cet Article."
       });
     });
 };
+
 // ::::::::::   Retrieve all Articles from the database  :::::::::::::::::::::::::::::::::::
 exports.findAll = (req, res) => {
   const title = req.query.title;
@@ -143,28 +139,57 @@ exports.addArticle = (userId, articleId) => {
 
 //:::::::::::   Delete an Article with id :::::::::::::
 exports.delete = (req, res) => {
+  const token = req.headers['x-access-token'];
+  const decoded = jwt.decode(token);
   const article_id = req.params.id;
 
-  Article.destroy({
-    where: { article_id: article_id },
-  })
-    .then((num) => {
-      if (num == 1) {
-        res.send({
-          message: "Article was deleted successfully!",
-        });
-      } else {
-        res.send({
-          message: `Cannot delete Article with id=${article_id}. Maybe Article was not found!`,
-        });
-      }
-    })
-    .catch((err) => {
-      res.status(500).send({
-        message: "Could not delete Article with id=" + article_id,
+  User.findByPk(decoded.id)
+    .then((user) => {
+      let currentUserRole = null;
+      user.getRoles().then(roles => {
+        for (let i = 0; i < roles.length; i++) {
+          if (roles[i].name === "admin") {
+            currentUserRole = roles[i].name
+          }
+        }
+        Article.findByPk(req.params.id)
+          .then((article) => {
+            if (!article) {
+              console.log("Article not found!");
+              return null;
+            }
+            if (article.user_id === decoded.id || currentUserRole === "admin") {
+              Article.destroy({
+                where: { article_id: article_id },
+              })
+                .then((num) => {
+                  if (num == 1) {
+                    res.send({
+                      message: "Article was deleted successfully!",
+                    });
+                  } else {
+                    res.send({
+                      message: `Cannot delete Article with id=${article_id}. Maybe Article was not found!`,
+                    });
+                  }
+                })
+                .catch((err) => {
+                  res.status(500).send({
+                    message: "Could not delete Article with id=" + article_id,
+                  });
+                });
+            }
+            else {
+              res.status(403).send({
+                message: "You don't have permission to perform this action",
+              });
+            }
+          });
       });
-    });
+    })
 };
+
+
 
 // Delete all Articles from the database.
 exports.deleteAll = (req, res) => {
@@ -182,19 +207,4 @@ exports.deleteAll = (req, res) => {
       });
     });
 };
-
-// Find all published Articles
-exports.findAllPublished = (req, res) => {
-  Article.findAll({ where: { published: true } })
-    .then((data) => {
-      res.send(data);
-    })
-    .catch((err) => {
-      res.status(500).send({
-        message:
-          err.message || "Some error occurred while retrieving Articles.",
-      });
-    });
-};
-
